Split the widget session specs and share the chat-service stubs
